changed delete

+++ backend/controllers/product.controller.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import Product from "../models/product.model.js";
 export const getProducts = async (req,res) =>{
     try {
@@ -47,6 +48,11 @@ export const createProduct = async (req, res)=>{
 
 export const deleteProduct = async(req,res)=>{
     const {id} = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({ success: false, message: "Invalid product Id"});
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "product Deleted"});
